+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/devices/DeviceCoordinator.java
@@ -201,8 +201,6 @@ public interface DeviceCoordinator {
      */
     boolean supportsHeartRateMeasurement(GBDevice device);
 
-    int getTapString();
-
     /**
      * Returns the readable name of the manufacturer.
      */
@@ -242,4 +240,9 @@ public interface DeviceCoordinator {
      */
     boolean supportsCalendarEvents();
 
+    /**
+     * Indicates whether the device supports getting a stream of live data.
+     * This can be live HR, steps etc.
+     */
+    boolean supportsRealtimeData();
 }